Disable unwanted and broken modules
