+++ disabled_modules/rice.py
