+++ disabled_modules/stickers.py
